server/captured-shop-server/auth-server/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
astroid==3.0.0
bcrypt==4.0.1
click==8.1.7
dill==0.3.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.11.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pylint==3.0.0
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
fastapi-mail==1.4.1
aiomysql==0.2.0
slack_sdk==3.26.2
greenlet==3.1.1